Amairul Hassan
“College Management System” has been created using two mastermind Language JAVA
& ORACLE. The former is being in the foreground and the latter in the background
respectively.

The background language is favorite of many programmers because it offers the much
wanted structure query facilities. Oracle provides high securities at the database are stored
in the logical area called “db Engine”. As db engine is created in the virtual memory and
end user cannot access that directly it. Now a day’s Java is very usable programming
language because it is purely object Oriented Language and java program’s runs on any
machine easily. With these thought in mind “College Management System” has been
develop after brain storming the needs and requirements of the College.